+++ src/oracle-database/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+2.0.4
+++++++
+* Fixed autonomous database update discriminator databaseType
+* Save generated autonomous database wallets as zip files
+* Fixed autonomous database backup create and show result lookup for OCI-backed backups
+
 2.0.3
 ++++++
 * Fixed autonomous database  discriminator databaseType in cross region DR related commands